Linux file mapping

+++ SystemResource/Source/Media/Sound/OGG/OGGPage.h
@@ -8,7 +8,7 @@ namespace BF
 	{
 		unsigned char Version; // Often a zero
 		unsigned char HeaderType;
-		size_t GranulePosition;
+		unsigned long long GranulePosition; // 8 Bytes
 		unsigned int SerialNumber;
 		unsigned int SequenceNumber;
 		unsigned int CRC32CheckSum; // CRC32, is generated using a polynomial value of 0x04C11DB7.
